apps/wdg-agents/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "apps/wdg_agents" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../../coverage/apps/wdg_agents/html' --cov-report xml:'../../../../coverage/apps/wdg_agents/coverage.xml' --html='../../../reports/apps/wdg_agents/unittests/html/index.html' --junitxml='../../../reports/apps/wdg_agents/unittests/junit.xml'"

[tool.poetry]
name = "wdg-agents"
version = "1.0.0"
description = "RAG Agents for WDG Project."
authors = [ "Andrew <andrew@andrew-k.us>" ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "wdg_agents"
  from = "apps"

  [tool.poetry.dependencies]
  python = ">=3.9,<4"
  langchain-google-genai = "^2.0.0"
  langchain-community = "^0.3.0"
  pypdf = "^4.3.1"
  chromadb = "^0.5.5"
  langchain-ollama = "^0.2.0"
  langchain-chroma = "^0.1.4"
  sentence-transformers = "^3.2.1"
  flask = "^3.1.0"
  flask-cors = "^5.0.0"
  python-dotenv = "^1.0.1"
  langchain-google-community = "^2.0.3"
  langchain-google-vertexai = "^2.0.10"
  langgraph = "^0.2.61"
  langchain-experimental = "^0.3.4"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
agent = 'wdg_agents.agent_1:main'

example = 'wdg_agents.cli:main'

start-flask = 'wdg_agents.api:run_server'
